requirements.txt
flask
flask-sqlalchemy
flask-login
flask-mail
flask-migrate
flask-wtf
wtforms
gunicorn
plotly
numpy
pandas
python-dotenv